# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "EditorPreference": {
        "properties": {
          "aiViewMode": {
            "default": "visual",
            "enum": [
              "visual",
              "text"
            ],
            "title": "Aiviewmode",
            "type": "string"
          },
          "userViewMode": {
            "default": "visual",
            "enum": [
              "visual",
              "text"
            ],
            "title": "Userviewmode",
            "type": "string"
          }
        },
        "title": "EditorPreference",
        "type": "object"
      },
      "EditorPreferencePatch": {
        "properties": {
          "aiViewMode": {
            "default": "visual",
            "enum": [
              "visual",
              "text"
            ],
            "title": "Aiviewmode",
            "type": "string"
          },
          "userViewMode": {
            "default": "visual",
            "enum": [
              "visual",
              "text"
            ],
            "title": "Userviewmode",
            "type": "string"
          }
        },
        "title": "EditorPreferencePatch",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ProjectPatchRequest": {
        "properties": {
          "cover": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProjectCoverSelection"
              },
              {
                "type": "null"
              }
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "useDescriptionForAiPrompt": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Usedescriptionforaiprompt"
          }
        },
        "title": "ProjectPatchRequest",
        "type": "object"
      },
...
      "ProjectSummary": {
        "properties": {
          "assetCount": {
            "title": "Assetcount",
            "type": "integer"
          },
          "completedTaskCount": {
            "default": 0,
            "title": "Completedtaskcount",
            "type": "integer"
          },
          "coverUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coverurl"
          },
          "createdAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "generationWarnings": {
            "items": {
              "type": "string"
            },
            "title": "Generationwarnings",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "newResults": {
            "items": {
              "additionalProperties": {
                "type": "string"
              },
              "type": "object"
            },
            "title": "Newresults",
            "type": "array"
          },
          "status": {
            "enum": [
              "idle",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "taskCount": {
            "title": "Taskcount",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "taskCount",
          "assetCount",
          "updatedAt"
        ],
        "title": "ProjectSummary",
        "type": "object"
      },
...
      "PromptEnhancementPreviewRequest": {
        "properties": {
          "context": {
            "$ref": "#/components/schemas/PromptEnhancementContextRequest",
            "default": {
              "includePreviousTaskSummary": false,
              "includeProjectBackground": false
            }
          },
          "generation": {
            "$ref": "#/components/schemas/PromptEnhancementGenerationRequest"
          },
          "media": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/PromptEnhancementMediaRequest"
            },
            "title": "Media",
            "type": "array"
          },
          "previousTaskId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previoustaskid"
          },
          "target": {
            "title": "Target",
            "type": "string"
          },
          "userPrompt": {
            "title": "Userprompt",
            "type": "string"
          }
        },
        "required": [
          "target",
          "userPrompt",
          "generation"
        ],
        "title": "PromptEnhancementPreviewRequest",
        "type": "object"
      },
      "PromptEnhancementPreviewView": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "modelId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modelid"
          },
          "previewId": {
            "title": "Previewid",
            "type": "string"
          },
          "prompt": {
            "title": "Prompt",
            "type": "string"
          },
          "providerId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Providerid"
          },
          "skillVersion": {
            "title": "Skillversion",
            "type": "string"
          },
          "targetSkill": {
            "title": "Targetskill",
            "type": "string"
          }
        },
        "required": [
          "previewId",
          "createdAt",
          "prompt",
          "targetSkill",
          "skillVersion",
          "providerId",
          "modelId"
        ],
        "title": "PromptEnhancementPreviewView",
        "type": "object"
      },
...
      "TaskReorderRequest": {
        "properties": {
          "taskIds": {
            "items": {
              "type": "string"
            },
            "title": "Taskids",
            "type": "array"
          }
        },
        "required": [
          "taskIds"
        ],
        "title": "TaskReorderRequest",
        "type": "object"
      },
      "TaskSaveRequest": {
        "properties": {
          "aiEnhancedPrompt": {
            "default": "",
            "title": "Aienhancedprompt",
            "type": "string"
          },
          "assetBindings": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/TaskAssetRef"
            },
            "title": "Assetbindings",
            "type": "array"
          },
          "durationSeconds": {
            "default": 6.0,
            "exclusiveMinimum": 0.0,
            "title": "Durationseconds",
            "type": "number"
          },
          "editorPreference": {
            "$ref": "#/components/schemas/EditorPreference",
            "default": {
              "aiViewMode": "visual",
              "userViewMode": "visual"
            }
          },
          "generation": {
            "$ref": "#/components/schemas/GenerationSettings",
            "default": {
              "contextMode": "尾帧承接",
              "mode": "全能参考",
              "quality": "标准",
              "resolution": "1080p"
            }
          },
          "promptSource": {
            "default": "user",
            "enum": [
              "user",
              "ai"
            ],
            "title": "Promptsource",
            "type": "string"
          },
          "revision": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision"
          },
          "scriptSource": {
            "default": "",
            "title": "Scriptsource",
            "type": "string"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "userIntent": {
            "default": "",
            "title": "Userintent",
            "type": "string"
          },
          "userPrompt": {
            "default": "",
            "title": "Userprompt",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "TaskSaveRequest",
        "type": "object"
      },
      "TaskSummary": {
        "properties": {
          "assetCount": {
            "title": "Assetcount",
            "type": "integer"
          },
          "displayNumber": {
            "title": "Displaynumber",
            "type": "integer"
          },
          "durationSeconds": {
            "title": "Durationseconds",
            "type": "number"
          },
          "generationStatusNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generationstatusnote"
          },
          "generationSummary": {
            "$ref": "#/components/schemas/GenerationSummary"
          },
          "generationWarnings": {
            "items": {
              "type": "string"
            },
            "title": "Generationwarnings",
            "type": "array"
          },
          "hasActivePromptJob": {
            "default": false,
            "title": "Hasactivepromptjob",
            "type": "boolean"
          },
          "hasActiveVideoJob": {
            "default": false,
            "title": "Hasactivevideojob",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "latestPromptRevisionId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestpromptrevisionid"
          },
          "latestVideoResultId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latestvideoresultid"
          },
          "previewUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previewurl"
          },
          "primaryResult": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PrimaryResultView"
              },
              {
                "type": "null"
              }
            ]
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "promptEnhancementStatus": {
            "default": "idle",
            "title": "Promptenhancementstatus",
            "type": "string"
          },
          "promptExcerpt": {
            "title": "Promptexcerpt",
            "type": "string"
          },
          "promptReviewStatus": {
            "default": "not_ready",
            "enum": [
              "not_ready",
              "pending_review",
              "approved"
            ],
            "title": "Promptreviewstatus",
            "type": "string"
          },
          "resultCount": {
            "title": "Resultcount",
            "type": "integer"
          },
          "status": {
            "enum": [
              "idle",
              "running",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "timing": {
            "$ref": "#/components/schemas/TaskTiming"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "videoGenerationStatus": {
            "default": "idle",
            "title": "Videogenerationstatus",
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayNumber",
          "title",
          "promptExcerpt",
          "previewUrl",
          "status",
          "progress",
          "assetCount",
          "resultCount",
          "durationSeconds",
          "generationSummary"
        ],
        "title": "TaskSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "ShotMill",
    "version": "0.3"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/projects/{project_id}": {
      "patch": {
        "operationId": "update_project_api_v1_projects__project_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectPatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Project",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/prompt-enhancement-previews": {
      "post": {
        "operationId": "preview_prompt_enhancement_api_v1_projects__project_id__prompt_enhancement_previews_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptEnhancementPreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptEnhancementPreviewView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Preview Prompt Enhancement",
        "tags": [
          "prompt-enhancement"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/tasks": {
      "post": {
        "operationId": "create_task_api_v1_projects__project_id__tasks_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/projects/{project_id}/tasks/reorder": {
      "post": {
        "operationId": "reorder_tasks_api_v1_projects__project_id__tasks_reorder_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskReorderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Reorder Tasks",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/projects/{project_id}/tasks/{task_id}": {
      "patch": {
        "operationId": "update_task_api_v1_projects__project_id__tasks__task_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskSaveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Task",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/projects/{project_id}/tasks/{task_id}/editor": {
      "get": {
        "operationId": "get_task_editor_api_v1_projects__project_id__tasks__task_id__editor_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskEditorView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task Editor",
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/projects/{project_id}/tasks/{task_id}/editor-preference": {
      "patch": {
        "operationId": "update_editor_preference_api_v1_projects__project_id__tasks__task_id__editor_preference_patch",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditorPreferencePatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditorPreference"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Editor Preference",
        "tags": [
          "tasks"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/tasks/{task_id}/prompt-enhancements": {
      "post": {
        "operationId": "enhance_prompt_api_v1_projects__project_id__tasks__task_id__prompt_enhancements_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptEnhancementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiPromptRevisionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Enhance Prompt",
        "tags": [
          "prompt-enhancement"
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/tasks/{task_id}/prompt-revisions": {
      "get": {
        "operationId": "list_prompt_revisions_api_v1_projects__project_id__tasks__task_id__prompt_revisions_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PromptRevisionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Prompt Revisions",
        "tags": [
          "prompt-enhancement"
        ]
      }
    },
    "/api/v1/projects/{project_id}/tasks/{task_id}/prompt-revisions/{revision_id}/select": {
      "post": {
        "operationId": "select_prompt_revision_api_v1_projects__project_id__tasks__task_id__prompt_revisions__revision_id__select_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision_id",
            "required": true,
            "schema": {
              "title": "Revision Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AiPromptRevisionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Prompt Revision",
        "tags": [
          "prompt-enhancement"
        ]
      }
    },
...
  }
}